Extract db connection into module
