+++ components/case/caseDao.js
@@ -1,40 +1,15 @@
 'use strict';
 
-const pgp = require('pg-promise');
-const connectionConfig = {
-  host: process.env.DB_HOST,
-  port: process.env.DB_PORT,
-  database: process.env.DB_NAME,
-  user: process.env.DB_USER,
-  password: process.env.DB_PW,
-};
-const pgOptions = {
-  receive: (e) => {
-    camelizeColumns(e.data);
-  },
-};
-const camelizeColumns = (data) => {
-  const template = data[0];
-  for (let prop in template) {
-    const camel = pgp.utils.camelize(prop);
-    if (!(camel in template)) {
-      for (let i = 0; i < data.length; i++) {
-        let d = data[i];
-        d[camel] = d[prop];
-        delete d[prop];
-      }
-    }
-  }
-};
-const pgPackage = pgp(pgOptions);
-const pg = pgPackage(connectionConfig);
+const dbUtils = require('../../utils/pgUtils');
+const db = dbUtils.db;
+const pgp = dbUtils.pgPkg;
 
-const caseCols = new pgPackage.helpers.ColumnSet(
+const caseCols = new pgp.helpers.ColumnSet(
   ['id', 'case_identifier', 'requisition_id', 'lims_ids'],
   { table: 'cardea_case' }
 );
 
-const archiveCols = new pgPackage.helpers.ColumnSet(
+const archiveCols = new pgp.helpers.ColumnSet(
   [
     'id',
     'created',
@@ -54,7 +29,7 @@ const archiveCols = new pgPackage.helpers.ColumnSet(
 
 const addCase = (cardeaCase) => {
   return new Promise((resolve, reject) => {
-    pg.task('add-case', (tx) => {
+    db.task('add-case', (tx) => {
       let query =
         pgp.helpers.insert(cardeaCase, caseCols) +
         ' ON CONFLICT (case_identifier) DO UPDATE SET ' +
@@ -78,7 +53,7 @@ const allCaseArchiveDataQuery =
 const getByCaseIdentifier = (caseIdentifier) => {
   let query = allCaseArchiveDataQuery + ' WHERE case_identifier = $1';
   return new Promise((resolve, reject) => {
-    pg.one(query, caseIdentifier)
+    db.one(query, caseIdentifier)
       .then((data) => {
         resolve(data);
       })
